% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkGeneCountDistribution.R
\name{checkCountDistributionPerGroup}
\alias{checkCountDistributionPerGroup}
\title{fit different count models}
\usage{
checkCountDistributionPerGroup(count, groupID, countPerCell,
  covariates = NULL, covariatesString = "1", largestKRemoved = 0)
}
\arguments{
\item{count}{the vector of counts}

\item{groupID}{the group ID character}

\item{countPerCell}{the total UMI per cell}

\item{covariates}{the covariates to account for}

\item{covariatesString}{the covariates string used in the formula}

\item{largestKRemoved}{remove the numbers correspond to the largest K}
}
\value{
a data frame with fitted information
}
\description{
This function fits the count with different count models including Poisson,
negative binomial and zero inflated negative binomial models
}
\examples{
\dontrun{
count = rnbinom(100, size = 2, mu = 0.1)
groupID = 1
countPerCell = floor(rnorm(100, 1e4, 3000))
fitResult = checkCountDistributionPerGroup(count, groupID, countPerCell)
print(t(fitResult))
}

}
