% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkGeneCountDistribution.R
\name{checkDataDistribution}
\alias{checkDataDistribution}
\title{fit different count models for a data matrix}
\usage{
checkDataDistribution(data, groups, countPerCell = NULL,
  covariates = NULL, largestKRemoved = 0, ncore = 1)
}
\arguments{
\item{data}{the gene-cell data matrix, it assumes the row names of the matrix
are the gene names}

\item{groups}{the group ID vector. If there are multiple groups, the results
will be put together with cbind}

\item{countPerCell}{the total UMI per cell. If it is NULL, it is calculated
by summing over all the gene counts for each cell.}

\item{covariates}{the covariates to account for}

\item{largestKRemoved}{remove the numbers correspond to the largest K}

\item{ncore}{the number of cores to use for parallel running}
}
\value{
a data frame with fitted information
}
\description{
This function fits the count with different count models including Poisson,
negative binomial and zero inflated negative binomial models. It invokes
checkCountDistributionPerGroup for the real fitting
}
\examples{
\dontrun{
gene1 = rnbinom(100, size = 0.1, mu = 0.1)
gene2 = rpois(100, lambda = 0.1)
data = rbind(gene1, gene2)
groups = rep(1, 100)
countPerCell = rep(1e4, 100)
fitResult = checkDataDistribution(data, groups, countPerCell)
print(t(fitResult))
}

}
