% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBID.R
\name{DEUsingNBID}
\alias{DEUsingNBID}
\title{Differential expression analysis using NBID on a data matrix}
\usage{
DEUsingNBID(data, groups, covariates = NULL, countPerCell = NULL,
  sizeFactor = NULL, singleDispersion = F, dispMethod = "poisson-ML",
  ncore = 1)
}
\arguments{
\item{data}{the gene-cell matrix}

\item{groups}{the group vector}

\item{covariates}{the covariates}

\item{countPerCell}{the total UMI per cell}

\item{sizeFactor}{the normalization factor, the effective count used will be
the countPerCell * sizeFactor. The size factor from
package scran needs to be divided by countPerCell first,
i.e., the size factor from scran is used as the effective
count directly.}

\item{singleDispersion}{whether to use one single dispersion for all cells.
All groups will share this single dispersion parameter, i.e., no
independent dispersions are used}

\item{dispMethod}{the method to estimate dispersions}

\item{ncore}{the number of cores to use for parallel running

This function invokes NBID for each row of the data matrix.}
}
\value{
a matrix of columns as follows:
pvalue: p value for each gene
LR: likelihood ratio test statistic
beta: the coefficient
dispersionGroup1, dispersionGroup2: estimated dispersions
log2FC: log2 fold change
}
\description{
Differential expression analysis using NBID on a data matrix
}
\examples{
\dontrun{
data(smallData)
result = DEUsingNBID(smallData$count, smallData$groupLabel)
head(result)
}

}
