% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodnessOfFit.R
\name{downsampleData}
\alias{downsampleData}
\title{downsample a data matrix}
\usage{
downsampleData(data, groupLabel, newTotalCounts = NULL,
  newQuantile = 0.25)
}
\arguments{
\item{data}{the gene*cell count matrix}

\item{groupLabel}{the label of each cell}

\item{newTotalCounts}{the UMI count to after downsampling}

\item{newQuantile}{the common UMI after downsampling is set by the quantile
of the total UMI across cells before downsampling
If newTotalCounts is set, this parameter will be ignored}
}
\value{
A list of two components
        data: this is the count matrix
        groupLable: this is the group label of cells
}
\description{
downsample a data matrix
}
