% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodnessOfFit.R
\name{goodnessOfFit}
\alias{goodnessOfFit}
\title{goodness of fit for a count vector}
\usage{
goodnessOfFit(count, distribution, plot = F)
}
\arguments{
\item{count}{a count vector}

\item{distribution}{"poisson" for Poisson distribution,
"nb" for the negative binomial distribution, "zinb" for the zero
inflated negative binomial distribution}

\item{plot}{If TRUE, plot the empirical and theoretical density and
cumulative distribution function using package fitdistrplus}
}
\value{
a data frame with the pvalue of the goodness of fit and whether the
        fit is converged
}
\description{
The distribution can be Poisson, negetive binomial or zero inflated
negative binomial (ZINB). For Poisson, glm is used for the fitting
For NB, glm.nb is used for the fitting. For ZINB,
a NB is first fitted and then zeroinfl is used for fitting
}
\examples{
\dontrun{
gene1 = rnbinom(500, size = 0.1, mu = 1)
print(goodnessOfFit(gene1, "poisson"))
print(goodnessOfFit(gene1, "nb"))
print(goodnessOfFit(gene1, "zinb"))

gene2 = rpois(500, lambda = 1)
print(goodnessOfFit(gene2, "poisson"))
print(goodnessOfFit(gene2, "nb"))
print(goodnessOfFit(gene2, "zinb"))
}

}
