% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodnessOfFit.R
\name{goodnessOfFitOnData}
\alias{goodnessOfFitOnData}
\title{Run goodness of fit on a count matrix. It first downsamples all cells to a
common total count. Cells with total count less than this threshold will
be discarded}
\usage{
goodnessOfFitOnData(countMatrix, dist = c("nb"),
  downsampleCount = NULL, downsampleQuantile = 0.1, ncore = 1)
}
\arguments{
\item{countMatrix}{a gene by cell matrix}

\item{dist}{a vector of distributions from "poisson", "nb", "zinb"}

\item{downsampleCount}{the common total count per cell used for downsampling}

\item{downsampleQuantile}{the quantile of the total count per cell used for
downsampling. If downsampleCount is set, this will be ignored.
If both are set to NULL, then no downsampling is performed}

\item{ncore}{the number of cores to use for parallel running}
}
\description{
Run goodness of fit on a count matrix. It first downsamples all cells to a
common total count. Cells with total count less than this threshold will
be discarded
}
\examples{
\dontrun{
load(smallData)
index1 = (smallData$groupLabel == 1)
data1 = smallData$count[, index1]
group1 = smallData$groupLabel[index1]
goodnessResult = goodnessOfFitOnData(data1)
head(goodnessResult)
}

}
