% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBID.R
\name{NBID}
\alias{NBID}
\title{negative binomial models allowing independent dispersions}
\usage{
NBID(count, groups, countPerCell, covariates, sizeFactor = NULL,
  singleDispersion = F, dispMethod = "poisson-ML")
}
\arguments{
\item{count}{the vector of gene counts}

\item{groups}{the vector of group information}

\item{covariates}{the covariaites}

\item{sizeFactor}{the normalization factor, the effective count used will be
the countPerCell * sizeFactor. The size factor from
package scran needs to be divided by countPerCell first,
i.e., the size factor from scran is used as the effective
count directly.}

\item{singleDispersion}{whether to use one single dispersion for all cells.
All groups will share this single dispersion parameter, i.e., no
independent dispersions are used}

\item{dispMethod}{the method to estimate the dispersion}

\item{countPercell}{the total UMI of each cell}
}
\value{
a vector of the following:

         LR: likelihood rato

         beta: the effect parameter

         dispersionOutput: dispersions of different groups
}
\description{
This function tests the following hypothesis
H0: the same mu for different groups
H1: different mu for different groups
In both cases, the dispersion is group specific. The dispersion is estimated
only for the full model and then used for the reduced model
}
