% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkGeneCountDistribution.R
\name{summaryComparison}
\alias{summaryComparison}
\title{summarize the model comparison results}
\usage{
summaryComparison(result)
}
\arguments{
\item{result}{}
}
\value{
a list of two components:
     modelCount: summarized count of best fitted models
     modelStats: the FDR and p-values of hypothesis testing results
}
\description{
This function summarizes the results and count the number of models that are
selected for each gene. The selection criterion can be hypothesis testing
based, or AIC based.
}
\examples{
\dontrun{
gene1 = rnbinom(100, size = 0.1, mu = 0.1)
gene2 = rpois(100, lambda = 0.1)
data = rbind(gene1, gene2)
groups = rep(1, 100)
countPerCell = rep(1e4, 100)
fitResult = checkDataDistribution(data, groups, countPerCell)
summaryResult = summaryComparison(fitResult)
print(summaryResult)
}

}
